requirements.txt
# Core ML libraries
torch>=1.13.0
torchvision>=0.14.0
transformers>=4.20.0
tokenizers>=0.13.0

# Scientific computing
numpy>=1.21.0
scipy>=1.9.0
scikit-learn>=1.1.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.6.0

# Data handling
pandas>=1.4.0
datasets>=2.0.0

# Configuration and logging
pyyaml>=6.0
hydra-core>=1.2.0
wandb>=0.13.0
tensorboard>=2.10.0

# Utilities
tqdm>=4.64.0
rich>=12.0.0
click>=8.0.0
pathlib2>=2.3.0

# Optional but recommended
faiss-cpu>=1.7.0  # For efficient similarity search
nltk>=3.7.0       # For text processing
spacy>=3.4.0      # For advanced NLP

# Development tools (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0 